+++ crates/testing-sdk/tests/sdk/rolling_top_k.rs
@@ -154,7 +154,7 @@ fn input_row(rn: u64, group: &str, window_start: u64, trader: u64, volume: f64) 
 			Value::float8(volume),
 		])
 		.with_fields(input_fields())
-		.with_time(DateTime::from_millis(window_start))
+		.with_time(DateTime::from_millis(window_start as i64))
 		.build()
 }
